+++ bot/modules/exceptions.py
@@ -52,12 +52,6 @@ class DatabaseError(Exception):
         message = "Error in user database: "+msg
         super().__init__(message)
 
-class StatusNotAllowed(Exception):
-    def __init__(self, name):
-        self.name = name
-        message = "This status is not allowed: "+name
-        super().__init__(message)
-
 class AccountsNotEnough(Exception):
     pass